Extract TokenRange
