+++ server/src/compiler_parser/nodes.ts
@@ -1,5 +1,5 @@
 import {TokenObject} from "../compiler_tokenizer/tokenObject";
-import {TextLocation} from "../compiler_tokenizer/textLocation";
+import {TokenRange} from "./tokenRange";
 
 export enum AccessModifier {
     Private = 'Private',
@@ -15,32 +15,6 @@ export enum TypeModifier {
 export enum ReferenceModifier {
     At = 'At',
     AtConst = 'AtConst',
-}
-
-export function getBoundingLocationBetween(start: TokenObject, end: TokenObject): TextLocation {
-    return start.location.withEnd(end.location.end);
-}
-
-export class TokenRange {
-    public constructor(
-        public readonly start: TokenObject,
-        public readonly end: TokenObject
-    ) {
-    }
-
-    /**
-     * Get text range covering two tokens
-     */
-    public getBoundingLocation(): TextLocation {
-        return getBoundingLocationBetween(this.start, this.end);
-    }
-
-    /**
-     *  Checks if the token spans a single line.
-     */
-    public isOneLine(): boolean {
-        return this.start.location.start.line === this.end.location.end.line;
-    }
 }
 
 export interface EntityAttribute {
